agentic-github-matcher/requirements.txt
# Agentic GitHub Matcher - Dependencies
# ======================================

# AutoGen - Agentic AI Framework (Classic version with ConversableAgent)
pyautogen==0.2.35

# LiteLLM - Model Gateway
litellm>=1.0.0

# Nemo Guardrails - Safety & Validation
nemoguardrails>=0.9.0

# Model Context Protocol (MCP) - GitHub Integration
# Official MCP Python SDK
mcp>=1.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.0.0

# HTTP Requests (for FastAPI/other services, NOT for GitHub - GitHub uses MCP)
requests>=2.31.0
aiohttp>=3.9.0

# OpenAI SDK
openai>=1.0.0

# Rich Console Output
rich>=13.0.0

# YAML Processing
pyyaml>=6.0.0

# Async Support
asyncio-throttle>=1.0.0

# FastAPI - Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

